sherlock: print reg

+++ sherlock/actions/kill.c
@@ -9,10 +9,16 @@
 
 #include "../action.h"
 #include <signal.h>
+#include <stdio.h>
 
 REG_ACTION(kill)
 {
-	pr_debug("action: kill");
-	kill(tracee->pid, SIGKILL);
-	RET_ACTION(tracee, TRACEE_KILLED);
+	pr_info_raw("Do you really want to kill the tracee (Y / N): ");
+	char opt[8];
+	fgets(opt, 8, stdin);
+	if (opt[0] == 'y' || opt[0] == 'Y') {
+		kill(tracee->pid, SIGKILL);
+		RET_ACTION(tracee, TRACEE_KILLED);
+	}
+	RET_ACTION(tracee, TRACEE_STOPPED);
 }
